fix: deduplicate admin check into shared module
