+++ turbo/apps/web/app/api/zero/agents/[id]/skills/route.ts
@@ -16,36 +16,12 @@ import { zeroSkills } from "../../../../../../src/db/schema/zero-skill";
 import { agentComposes } from "../../../../../../src/db/schema/agent-compose";
 import { eq, and, inArray } from "drizzle-orm";
 import { buildComposeContent } from "../../../../../../src/lib/zero/build-compose-content";
-import { isDefaultAgentCompose } from "../../../../../../src/lib/zero/resolve-default-agent";
+import { requireAdminForDefaultAgent } from "../../../../../../src/lib/zero/require-admin";
 import { uploadSkillServerSide } from "../../../../../../src/lib/storage/skill-upload";
 import { SEED_SKILLS } from "../../../../../../src/lib/zero/seed-skills";
 import { logger } from "../../../../../../src/lib/logger";
 
 const log = logger("api:zero-agents:skills");
-
-type ForbiddenResponse = {
-  status: 403;
-  body: { error: { message: string; code: string } };
-};
-
-async function requireAdminForDefaultAgent(
-  orgId: string,
-  composeId: string,
-  memberRole: string,
-): Promise<ForbiddenResponse | null> {
-  if (memberRole === "admin") return null;
-  const isDefault = await isDefaultAgentCompose(orgId, composeId);
-  if (!isDefault) return null;
-  return {
-    status: 403 as const,
-    body: {
-      error: {
-        message: "Only org admins can modify the default agent's skills",
-        code: "FORBIDDEN",
-      },
-    },
-  };
-}
 
 const router = tsr.router(zeroAgentSkillsCollectionContract, {
   list: async ({ params, headers }, { request }) => {
@@ -155,6 +131,7 @@ const router = tsr.router(zeroAgentSkillsCollectionContract, {
       org.orgId,
       existing.id,
       member.role,
+      "skills",
     );
     if (forbidden) return forbidden;
